+++ sdk/resourcemanager/devopsinfrastructure/armdevopsinfrastructure/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release History
 
+## 1.1.0 (2024-12-10)
+### Features Added
+
+- New enum type `AvailabilityStatus` with values `AvailabilityStatusAvailable`, `AvailabilityStatusUnavailable`
+- New enum type `CheckNameAvailabilityReason` with values `CheckNameAvailabilityReasonAlreadyExists`, `CheckNameAvailabilityReasonInvalid`
+- New enum type `ResourceType` with values `ResourceTypeMicrosoftDevOpsInfrastructurePools`
+- New function `*PoolsClient.CheckNameAvailability(context.Context, CheckNameAvailability, *PoolsClientCheckNameAvailabilityOptions) (PoolsClientCheckNameAvailabilityResponse, error)`
+- New struct `CheckNameAvailability`
+- New struct `CheckNameAvailabilityResult`
+
+
 ## 1.0.0 (2024-11-20)
 ### Breaking Changes